+++ ruuvi-gateway.py
@@ -1,8 +1,8 @@
 #----------------------------------------------------------------------------
 # Author  : Aaron Finney
 # Created : September 14, 2021
-# Updated : September 15, 2021
-# Version : 0.0.3
+# Updated : September 18, 2021
+# Version : 0.1.0
 # 
 # This script watches an MQTT topic for messages from a Ruuvi Gateway, decodes
 # them, performs basic processing (if configured), and re-emits them to MQTT
@@ -14,75 +14,6 @@ from ruuvi_decoders import Df3Decoder, Df5Decoder
 from json import dumps
 import ast
 import operator
-
-# Configuration variables
-MQTT_TOPIC_PREFIX = "ruuvi" # this must match what's configured in the Gateway MQTT settings
-DECIMAL_PLACES = 2 # number of decimal places for values
-BATTERY_RANGE = {"min":2500,"max":3000} # min/max battery range, used to calcuate battery level
-# RUUVITAGS defines the tags, by MAC address, that the script will process; any tag not included
-# in this dictionary will be ignored by the script. The value for each item is used as the string
-# to replace <TAG_NAME> in the emitted MQTT message payload
-RUUVITAGS = {"f33c5fc10393":"Living Room Ruuvitag",
-            "dc757bb86337":"Middle Bedroom Ruuvitag",
-            "f9fd6a322a01":"Front Bedroom Ruuvitag",
-            "dc2c241a4b24":"Kitchen Ruuvitag",
-            "ef336477a31a":"Back Bedroom Ruuvitag"}
-# MEASUREMENT CONFIG specifies which measurements to write back to MQTT, any transformations to be performed
-# on the values, and the config payload. The special values <MAC> and <TAG_NAME> are
-# replaced prior to emitting the config message to MQTT
-MEASUREMENT_CONFIG = {"temperature": 
-                        {"config":
-                            {"stat_t":"homeassistant/sensor/<MAC>/temperature/state",
-                            "json_attr_t":"homeassistant/sensor/<MAC>/temperature/attributes",
-                            "name":"<TAG_NAME> Temperature",
-                            "unit_of_meas":"\u00b0F",
-                            "dev_cla":"temperature",
-                            "uniq_id":"ruuvitag_<MAC>_temperature",
-                            "device":
-                                {"ids":"ruuvitag-<MAC>",
-                                "mf":"Ruuvi Innovations Ltd",
-                                "mdl": "RuuviTag",
-                                "name": "<TAG_NAME>"
-                                }
-                            },
-                        "transform": "(<VALUE>*(9/5))+32",
-                        "source_metric": "temperature"},
-                      "humidity": 
-                        {"config":
-                            {"stat_t":"homeassistant/sensor/<MAC>/humidity/state",
-                            "json_attr_t":"homeassistant/sensor/<MAC>/humidity/attributes",
-                            "name":"<TAG_NAME> Humidity",
-                            "unit_of_meas":"%",
-                            "dev_cla":"humidity",
-                            "uniq_id":"ruuvitag_<MAC>_humidity",
-                            "device":
-                                {"ids":"ruuvitag-<MAC>",
-                                "mf":"Ruuvi Innovations Ltd",
-                                "mdl": "RuuviTag",
-                                "name": "<TAG_NAME>"
-                                }
-                            },
-                        "transform": None,
-                        "source_metric": "humidity"},
-                      "battery_level":
-                        {"config":
-                            {"stat_t":"homeassistant/sensor/<MAC>/battery_level/state",
-                            "json_attr_t":"homeassistant/sensor/<MAC>/battery_level/attributes",
-                            "name":"<TAG_NAME> Battery Level",
-                            "unit_of_meas":"%",
-                            "dev_cla":"battery",
-                            "uniq_id":"ruuvitag_<MAC>_battery",
-                            "device":
-                                {"ids":"ruuvitag-<MAC>",
-                                "mf":"Ruuvi Innovations Ltd",
-                                "mdl": "RuuviTag",
-                                "name": "<TAG_NAME>"
-                                }
-                            },
-                        "transform": f"((<VALUE> - {BATTERY_RANGE['min']}) / ({BATTERY_RANGE['max']} - {BATTERY_RANGE['min']})) * 100",
-                        "source_metric": "battery"}
-                     }
-
 
 def sensor_exists(mac):
     try:
@@ -98,6 +29,8 @@ def eval_(node):
 
     if isinstance(node, ast.Num): # <number>
         return node.n
+    elif isinstance(node, ast.Constant): # <constant>
+        return node.n
     elif isinstance(node, ast.BinOp): # <left> <operator> <right>
         return operators[type(node.op)](eval_(node.left), eval_(node.right))
     elif isinstance(node, ast.UnaryOp): # <operator> <operand> e.g., -1
@@ -106,14 +39,17 @@ def eval_(node):
         raise TypeError(node)
 
 
-@mqtt_trigger(f"{MQTT_TOPIC_PREFIX}/#")
+@mqtt_trigger(f"{pyscript.app_config['mqtt_topic_prefix']}/#")
 def mqtt_msg(topic=None, payload=None, payload_obj=None):
+    ruuvitags = pyscript.app_config['ruuvitags']
+    measurement_config = pyscript.app_config['measurements']
+    decimal_places = pyscript.app_config['decimal_places']
     # get the MAC from the topic because data format 3 doesn't include it
     raw_mac = topic.split('/')[-1]
     mac = raw_mac.lower().replace(':','')
-    if mac not in RUUVITAGS:
+    if mac not in ruuvitags:
         # this sensor is not in the list of tags to process, ignore
-        log.warning(f"MAC addres {mac} not in RUUVITAGS, ignoring tag")
+        log.warning(f"MAC addres {mac} not in ruuvitags, ignoring tag")
         return
     if payload_obj.get("data") is not None:
         try:
@@ -130,20 +66,24 @@ def mqtt_msg(topic=None, payload=None, payload_obj=None):
                 decoder = Df5Decoder()
                 data = decoder.decode_data(clean_data)
 
-            for m_type in MEASUREMENT_CONFIG:
-                value = data.get(MEASUREMENT_CONFIG[m_type]["source_metric"])
-                if MEASUREMENT_CONFIG[m_type]["transform"] is not None:
-                    value = eval_(ast.parse(f"{MEASUREMENT_CONFIG[m_type]['transform']}".replace("<VALUE>",str(value)), mode='eval').body)
+            for m_type in measurement_config:
+                value = data.get(measurement_config[m_type]["source_metric"])
+                if measurement_config[m_type]["transform"] is not None:
+                    value = eval_(ast.parse(f"{measurement_config[m_type]['transform']}".replace("<VALUE>",str(value)), mode='eval').body)
                 if isinstance(value,float):
-                    value = round(value, DECIMAL_PLACES)
+                    value = round(value, decimal_places)
 
                 if not sensor_exists(mac):
                     config_topic = f"homeassistant/sensor/{mac}/{m_type}/config"
-                    config_payload = dumps(MEASUREMENT_CONFIG[m_type]['config']).replace('<MAC>',mac).replace('<TAG_NAME>', RUUVITAGS[mac])
+                    config_payload = dumps(measurement_config[m_type]['config']).replace('<MAC>',mac).replace('<TAG_NAME>', ruuvitags[mac])
+                    attributes_topic = f"homeassistant/sensor/{mac}/{m_type}/attributes"
+                    attributes_payload = dumps(measurement_config[m_type]['attributes']).replace('<MAC>',mac)
                     mqtt.publish(topic=config_topic, payload=config_payload, retain='True')
+                    mqtt.publish(topic=attributes_topic, payload=attributes_payload, retain='True')
 
                 state_topic = f"homeassistant/sensor/{mac}/{m_type}/state"
                 mqtt.publish(topic=state_topic, payload=value, retain='True')
 
-        except (AttributeError, ValueError, TypeError):
+        except (AttributeError, ValueError, TypeError) as err:
             log.warning(f'Error --> {payload}')
+            log.warning(f'Error code -> {err}')