+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.3] - 9/15/2021
+## [0.1.0] - 9/18/2021
+### Added
+- The script now creates an "attributes" topic in MQTT for each ruuvitag
+
+### Changed
+- All configuration has been moved to the main HA /config/configuration.yaml file (sample file added)
 ### Added
 - Log output with MAC address for beaconing tags that aren't in RUUUVITAG (rudimentary discovery)
 ## [0.0.2] - 9/15/2021